refactor out shared code between steps
